Moved around code into diffrent files. Changed how menu is cleared. Menu close button is fixed. Saving an event is fixed.
