+++ src/functions.js
@@ -1,0 +1,70 @@
+import {Event, Menus} from './objects'
+
+/**
+ * Saves calender information to local storage
+ */
+export function save() {
+    //save events
+    let evs = {};
+    events.forEach((event, i) => {
+        evs[i] = event;
+    });
+
+    localStorage.setItem('events', JSON.stringify(evs));
+
+    //save stickers
+    let sckers = {};
+    stickers.forEach(function (sticker, i) {
+        sckers[i] = sticker;
+    });
+    localStorage.setItem('stickers', JSON.stringify(sckers));
+
+    //save frame
+    localStorage.setItem('frame', JSON.stringify(frame));
+}
+
+/**
+ * Loads and reinstates calender information from storage a
+ */
+export function load() {
+    let evs = JSON.parse(localStorage.getItem('events'));
+    let sckers = JSON.parse(localStorage.getItem('stickers'));
+    //let fme = JSON.parse(localStorage.getItem('frame'));
+
+    for (let e in evs) {
+        //add events to stored events
+        let newEvent = new Event(evs[e].time, evs[e].title, evs[e].notes, evs[e].tdID); //create a new event from the stored one
+        events.push(newEvent);
+        //parse event into an element and append to corresponding div
+        let td = document.getElementById(`${newEvent.tdID}`);
+        if (td) //check that query selection is not null
+            td.getElementsByClassName("eventContainer")[0].appendChild(newEvent.toElement());
+    }
+    //TODO: load stickers
+    for (let s in sckers) {
+        stickers.push(sckers[s]);
+    }
+    //frame = fme;
+}
+
+/**
+ * Fills in view menu
+ * @param {string} title title of event
+ * @param {string} time time of event
+ * @param {string} notes notes about event
+ */
+export function fillInView() {
+    let title = "(No Title)",
+        time = "(No Time)",
+        notes = "(No Notes)";
+
+    events.some(e => {
+        if (e.id == selectedEvent.id) {
+            if (e.title != "") title = e.title;
+            if (e.time != "") time = e.time;
+            if (e.notes != "") notes = e.notes;
+            return true;
+        }
+    })
+    Menus.contentGenerators.viewEvent(title, time, notes);
+}
